requirements/base.txt
PyQt6==6.4.2
darkdetect
attrs
cattrs
Pygments
dataset
mistune
jsonpath-ng
Faker